requirements.txt
amqp==5.3.1 ; python_version >= "3.10" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==5.0.1 ; python_version >= "3.10" and python_full_version < "3.11.3"
attrs==25.1.0 ; python_version >= "3.10" and python_version < "4.0"
billiard==4.2.1 ; python_version >= "3.10" and python_version < "4.0"
brotli==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
celery==5.4.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.10" and python_version < "4.0"
click-didyoumean==0.3.1 ; python_version >= "3.10" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.10" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
cron-descriptor==1.4.5 ; python_version >= "3.10" and python_version < "4.0"
dj-database-url==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
dj-rest-auth==3.0.0 ; python_version >= "3.10" and python_version < "4.0"
django-celery-beat==2.7.0 ; python_version >= "3.10" and python_version < "4.0"
django-celery-results==2.5.1 ; python_version >= "3.10" and python_version < "4.0"
django-cors-headers==4.7.0 ; python_version >= "3.10" and python_version < "4.0"
django-environ==0.11.2 ; python_version >= "3.10" and python_version < "4"
django-filter==24.3 ; python_version >= "3.10" and python_version < "4.0"
django-redis==5.4.0 ; python_version >= "3.10" and python_version < "4.0"
django-rest-knox==5.0.2 ; python_version >= "3.10" and python_version < "4.0"
django-timezone-field==7.1 ; python_version >= "3.10" and python_version < "4.0"
django==5.1.6 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.15.2 ; python_version >= "3.10" and python_version < "4.0"
drf-spectacular==0.28.0 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema-specifications==2024.10.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema==4.23.0 ; python_version >= "3.10" and python_version < "4.0"
kombu==5.4.2 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.10" and python_version < "4.0"
prompt-toolkit==3.0.50 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.10" and python_version < "4.0"
psycopg2==2.9.10 ; python_version >= "3.10" and python_version < "4.0"
python-crontab==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
python-decouple==3.8 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0"
redis==5.2.1 ; python_version >= "3.10" and python_version < "4.0"
referencing==0.36.2 ; python_version >= "3.10" and python_version < "4.0"
rpds-py==0.22.3 ; python_version >= "3.10" and python_version < "4.0"
sentry-sdk[django]==2.21.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2025.1 ; python_version >= "3.10" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.34.0 ; python_version >= "3.10" and python_version < "4.0"
vine==5.1.0 ; python_version >= "3.10" and python_version < "4.0"
wcwidth==0.2.13 ; python_version >= "3.10" and python_version < "4.0"
whitenoise[brotli]==6.9.0 ; python_version >= "3.10" and python_version < "4.0"
